Move all discovery code to 'core/discovery' package
