+++ core/discovery/discovery_test.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package registry
+package discovery
 
 import (
 	"sync"
